convnextv2_tiny_224/requirements_gs40.txt
# Base (root requirements)
setuptools>=65,<82
scanpy>=1.10
scikit-misc>=0.3
squidpy>=1.5
anndata>=0.10
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
scipy>=1.11
h5py>=3.9
pyarrow>=14.0
leidenalg>=0.10
celltypist>=1.6
shapely>=2.0
jupyter
ipywidgets
tqdm
tifffile>=2023.1
PyYAML>=6.0
Pillow>=10.0

# PyTorch (CUDA 11.8)
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118

# GS40 / CellViT Extras
opencv-python
zarr
openslide-python
ipykernel
